promote experimental tabnine inlay to default
