+++ src/main/java/com/tabnine/inline/render/TabnineInlay.kt
@@ -2,10 +2,6 @@ package com.tabnine.inline.render
 
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.editor.Editor
-import com.tabnine.capabilities.CapabilitiesService
-import com.tabnine.capabilities.Capability
-import com.tabnine.inline.render.experimental.ExperimentalTabnineInlay
-import com.tabnine.inline.render.preserved.DefaultTabnineInlay
 import com.tabnine.prediction.TabNineCompletion
 import java.awt.Rectangle
 
@@ -21,9 +17,7 @@ interface TabnineInlay {
     companion object {
         @JvmStatic
         fun create(): TabnineInlay {
-            val isAlpha = CapabilitiesService.getInstance().isCapabilityEnabled(Capability.ALPHA)
-
-            return if (isAlpha) ExperimentalTabnineInlay() else DefaultTabnineInlay()
+            return DefaultTabnineInlay()
         }
     }
 }
